requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
firebase-admin==7.1.0
pydantic==2.12.5
python-dotenv==1.2.1
redis==5.0.1

